Move micrscope back - reloc is confusing
