+++ software/control/gui_hcs.py
@@ -16,7 +16,7 @@ from control._def import *
 import control.widgets as widgets
 import pyqtgraph.dockarea as dock
 import squid.logging
-import squid.microscope
+import control.microscope
 
 log = squid.logging.get_logger(__name__)
 
@@ -107,7 +107,7 @@ class HighContentScreeningGui(QMainWindow):
 
         self.makeConnections()
 
-        self.microscope = squid.microscope.Microscope(self)
+        self.microscope = control.microscope.Microscope(self)
 
         # Move to cached position
         if HOMING_ENABLED_X and HOMING_ENABLED_Y and HOMING_ENABLED_Z:
