+++ software/control/microscope.py
